lib/python/requirements.txt
numpy>=1.26
psycopg[binary]>=3.1
umap-learn>=0.5
